improve tests and coverage

+++ pytradingbot/cores/properties.py
@@ -30,7 +30,7 @@ class PropertiesABC(ABC):
                  param: Dict[str, Any] = None):
 
         self.child = []
-        self.data = pd.Series()
+        self.data = pd.Series(dtype=float)
 
         self.parents: Dict[str, object] = {}
         if isinstance(parent, dict):
@@ -179,7 +179,7 @@ class AskLoad(Ask):
     """
     Ask value loaded
     """
-    def __init__(self, market=None, data: pd.Series = pd.Series()):
+    def __init__(self, market=None, data: pd.Series = pd.Series(dtype=float)):
         super().__init__(market=market)
         self.data = data
 
@@ -188,7 +188,7 @@ class BidLoad(Bid):
     """
     Bid value loaded
     """
-    def __init__(self, market=None, data: pd.Series = pd.Series()):
+    def __init__(self, market=None, data: pd.Series = pd.Series(dtype=float)):
         super().__init__(market=market)
         self.data = data
 
@@ -197,7 +197,7 @@ class VolumeLoad(Volume):
     """
     volume value of the market
     """
-    def __init__(self, market=None, data: pd.Series = pd.Series()):
+    def __init__(self, market=None, data: pd.Series = pd.Series(dtype=float)):
         super().__init__(market=market)
         self.data = data
